+++ LightBulb/Function/BinaryFunction.hpp
@@ -12,13 +12,13 @@ class BinaryFunction : public AbstractActivationFunction
 private:
 public:
 	// Returns 0 (input < threshold) or 1 (input > threshold)
-	float execute(float input, AbstractThreshold* threshold);
+	double execute(double input, AbstractThreshold* threshold);
 	// Error - No derivation available
-	float executeDerivation(float input, AbstractThreshold* threshold);
+	double executeDerivation(double input, AbstractThreshold* threshold);
 	// Inherited:
 	AbstractActivationFunction* getActivationFunctionCopy();
-	float getMaximum();
-	float getMinimum();
+	double getMaximum();
+	double getMinimum();
 	bool hasAMaxAndMinimum();
 };
