+++ LightBulb/Function/GaussianRBFFunction.hpp
@@ -14,11 +14,11 @@ class GaussianRBFFunction : public AbstractActivationFunction
 private:
 public:
 	// Inherited:
-	float execute(float input, AbstractThreshold* threshold);
-	float executeDerivation(float input, AbstractThreshold* threshold);
+	double execute(double input, AbstractThreshold* threshold);
+	double executeDerivation(double input, AbstractThreshold* threshold);
 	AbstractActivationFunction* getActivationFunctionCopy();
-	float getMaximum();
-	float getMinimum();
+	double getMaximum();
+	double getMinimum();
 	bool hasAMaxAndMinimum();
 };
